+++ axum-core/src/ext_traits/mod.rs
@@ -1,14 +1,33 @@
 pub(crate) mod request;
 pub(crate) mod request_parts;
-pub(crate) mod service;
 
 #[cfg(test)]
 mod tests {
     use std::convert::Infallible;
 
+    use crate::extract::{FromRef, FromRequestParts};
     use async_trait::async_trait;
-    use axum_core::extract::{FromRef, FromRequestParts};
     use http::request::Parts;
+
+    #[derive(Debug, Default, Clone, Copy)]
+    pub(crate) struct State<S>(pub(crate) S);
+
+    #[async_trait]
+    impl<OuterState, InnerState> FromRequestParts<OuterState> for State<InnerState>
+    where
+        InnerState: FromRef<OuterState>,
+        OuterState: Send + Sync,
+    {
+        type Rejection = Infallible;
+
+        async fn from_request_parts(
+            _parts: &mut Parts,
+            state: &OuterState,
+        ) -> Result<Self, Self::Rejection> {
+            let inner_state = InnerState::from_ref(state);
+            Ok(Self(inner_state))
+        }
+    }
 
     // some extractor that requires the state, such as `SignedCookieJar`
     pub(crate) struct RequiresState(pub(crate) String);
